requirements.txt
psutil
pypresence
pyobjc-framework-AppKit
pyobjc-framework-Quartz
pyobjc-framework-ApplicationServices

